+++ client/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app';
 
-import Providers from './provider';
 import { fontMono, fontSans } from '@/config/fonts';
+import Providers from '@/config/provider';
 import '@/styles/globals.css';
 
 export default function App({ Component, pageProps }: AppProps) {
